Adding initial hackadmin app
