+++ hackathon/urls.py
@@ -11,7 +11,6 @@ from .views import (
     view_hackathon,
     update_hackathon_status,
     change_awards,
-    hackathon_stats,
     judge_teams,
 )
 from teams.views import change_teams
@@ -34,7 +33,6 @@ urlpatterns = [
     path("<int:hackathon_id>/delete", delete_hackathon,
          name="delete_hackathon"),
     path('enroll/', enroll_toggle, name='enroll_toggle'),
-    path('stats/', hackathon_stats, name='hackathon_stats'),
     path('<int:hackathon_id>/judge_teams/', judge_teams,
         name="judge_teams"),
 ]
